requirements.txt
Flask
requests
Pillow
opencv-python-headless
numpy
